Make suggested imports its own diag + cleanup
