+++ server/locale/en-us/script.lua
@@ -2,6 +2,7 @@ DIAG_LINE_ONLY_SPACE    = 'Line with spaces only.'
 DIAG_LINE_POST_SPACE    = 'Line with postspace.'
 DIAG_UNUSED_LOCAL       = 'Unused local `{}`.'
 DIAG_UNDEF_GLOBAL       = 'Undefined global `{}`.'
+DIAG_SUGGESTED_IMPORT   = 'Potential Import for `{}`.'
 DIAG_UNDEF_FIELD        = 'Undefined field `{}`.'
 DIAG_UNDEF_ENV_CHILD    = 'Undefined variable `{}` (overloaded `_ENV` ).'
 DIAG_UNDEF_FENV_CHILD   = 'Undefined variable `{}` (inside module).'
